refactoring to group utils
