+++ novelsave/sources/source.py
@@ -1,6 +1,6 @@
 import re
 import time
-from typing import Tuple, List, Iterable, Union
+from typing import Tuple, List, Union
 from urllib.parse import urlparse
 
 import requests
@@ -8,7 +8,7 @@ from bs4 import BeautifulSoup, Comment
 from requests.cookies import RequestsCookieJar
 
 from ..models import Novel, Chapter
-from ..tools import StringTools
+from ..utils.helpers import StringHelper
 from ..exceptions import ResponseException
 
 header = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_1) AppleWebKit/537.36 (KHTML, like Gecko) '
@@ -145,14 +145,14 @@ class Source:
         """
         :return: suitable folder name from netloc
         """
-        return StringTools.slugify(urlparse(self.base).netloc, replace='_')
+        return StringHelper.slugify(urlparse(self.base).netloc, replace='_')
 
     def novel_folder_name(self, url):
         """
         :param url: novel url
         :return: suitable novel folder name
         """
-        return StringTools.slugify(url.strip('/ ').split('/')[-1])
+        return StringHelper.slugify(url.strip('/ ').split('/')[-1])
 
     # ---- Inspired from https://github.com/dipu-bd/lightnovel-crawler ----
     # ----      And almost a perfect copy of the functions below       ----
