+++ novelsave/utils/ui/figlet.py
@@ -1,4 +1,4 @@
-from ..meta import __version__
+from ...meta import __version__
 
 banner = r'''                              ___                                       
                              /\_ \
